+++ src/Silex/Provider/SecurityServiceProvider.php
@@ -15,6 +15,7 @@ use Silex\Application;
 use Silex\ServiceProviderInterface;
 use Symfony\Component\HttpFoundation\RequestMatcher;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Security;
 use Symfony\Component\Security\Core\SecurityContext;
 use Symfony\Component\Security\Core\SecurityContextInterface;
 use Symfony\Component\Security\Core\User\UserChecker;
@@ -27,6 +28,8 @@ use Symfony\Component\Security\Core\Authentication\AuthenticationProviderManager
 use Symfony\Component\Security\Core\Authentication\AuthenticationTrustResolver;
 use Symfony\Component\Security\Http\Authentication\DefaultAuthenticationSuccessHandler;
 use Symfony\Component\Security\Http\Authentication\DefaultAuthenticationFailureHandler;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
+use Symfony\Component\Security\Core\Authorization\AuthorizationChecker;
 use Symfony\Component\Security\Core\Authorization\Voter\RoleHierarchyVoter;
 use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManager;
@@ -72,9 +75,31 @@ class SecurityServiceProvider implements ServiceProviderInterface
         $app['security.access_rules'] = array();
         $app['security.hide_user_not_found'] = true;
 
-        $app['security'] = $app->share(function ($app) {
-            return new SecurityContext($app['security.authentication_manager'], $app['security.access_manager']);
-        });
+        $r = new \ReflectionMethod('Symfony\Component\Security\Http\Firewall\ContextListener', '__construct');
+        $params = $r->getParameters();
+        if ('Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface' === $params[0]->getClass()->getName()) {
+            $app['security.authorization_checker'] = $app->share(function ($app) {
+                return new AuthorizationChecker($app['security.token_storage'], $app['security.authentication_manager'], $app['security.access_manager']);
+            });
+
+            $app['security.token_storage'] = $app->share(function ($app) {
+                return new TokenStorage();
+            });
+
+            $app['security'] = $app->share(function ($app) {
+                // Deprecated, to be removed in 2.0
+                return new SecurityContext($app['security.token_storage'], $app['security.authorization_checker']);
+            });
+        } else {
+            $app['security.token_storage'] = $app['security.authorization_checker'] = $app->share(function ($app) {
+                return $app['security'];
+            });
+
+            $app['security'] = $app->share(function ($app) {
+                // Deprecated, to be removed in 2.0
+                return new SecurityContext($app['security.authentication_manager'], $app['security.access_manager']);
+            });
+        }
 
         $app['security.authentication_manager'] = $app->share(function ($app) {
             $manager = new AuthenticationProviderManager($app['security.authentication_providers']);
@@ -271,7 +296,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
 
         $app['security.access_listener'] = $app->share(function ($app) {
             return new AccessListener(
-                $app['security'],
+                $app['security.token_storage'],
                 $app['security.access_manager'],
                 $app['security.access_map'],
                 $app['security.authentication_manager'],
@@ -306,14 +331,19 @@ class SecurityServiceProvider implements ServiceProviderInterface
         });
 
         $app['security.last_error'] = $app->protect(function (Request $request) {
-            if ($request->attributes->has(SecurityContextInterface::AUTHENTICATION_ERROR)) {
-                return $request->attributes->get(SecurityContextInterface::AUTHENTICATION_ERROR)->getMessage();
+            if (class_exists('Symfony\Component\Security\Core\Security')) {
+                $error = Security::AUTHENTICATION_ERROR;
+            } else {
+                $error = SecurityContextInterface::AUTHENTICATION_ERROR;
+            }
+            if ($request->attributes->has($error)) {
+                return $request->attributes->get($error)->getMessage();
             }
 
             $session = $request->getSession();
-            if ($session && $session->has(SecurityContextInterface::AUTHENTICATION_ERROR)) {
-                $error = $session->get(SecurityContextInterface::AUTHENTICATION_ERROR)->getMessage();
-                $session->remove(SecurityContextInterface::AUTHENTICATION_ERROR);
+            if ($session && $session->has($error)) {
+                $error = $session->get($error)->getMessage();
+                $session->remove($error);
 
                 return $error;
             }
@@ -324,7 +354,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
         $app['security.context_listener._proto'] = $app->protect(function ($providerKey, $userProviders) use ($app) {
             return $app->share(function () use ($app, $userProviders, $providerKey) {
                 return new ContextListener(
-                    $app['security'],
+                    $app['security.token_storage'],
                     $userProviders,
                     $providerKey,
                     $app['logger'],
@@ -347,7 +377,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
         $app['security.exception_listener._proto'] = $app->protect(function ($entryPoint, $name) use ($app) {
             return $app->share(function () use ($app, $entryPoint, $name) {
                 return new ExceptionListener(
-                    $app['security'],
+                    $app['security.token_storage'],
                     $app['security.trust_resolver'],
                     $app['security.http_utils'],
                     $name,
@@ -401,7 +431,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
                 }
 
                 return new $class(
-                    $app['security'],
+                    $app['security.token_storage'],
                     $app['security.authentication_manager'],
                     isset($app['security.session_strategy.'.$name]) ? $app['security.session_strategy.'.$name] : $app['security.session_strategy'],
                     $app['security.http_utils'],
@@ -419,7 +449,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
         $app['security.authentication_listener.http._proto'] = $app->protect(function ($providerKey, $options) use ($app) {
             return $app->share(function () use ($app, $providerKey, $options) {
                 return new BasicAuthenticationListener(
-                    $app['security'],
+                    $app['security.token_storage'],
                     $app['security.authentication_manager'],
                     $providerKey,
                     $app['security.entry_point.'.$providerKey.'.http'],
@@ -431,7 +461,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
         $app['security.authentication_listener.anonymous._proto'] = $app->protect(function ($providerKey, $options) use ($app) {
             return $app->share(function () use ($app, $providerKey, $options) {
                 return new AnonymousAuthenticationListener(
-                    $app['security'],
+                    $app['security.token_storage'],
                     $providerKey,
                     $app['logger']
                 );
@@ -460,7 +490,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
                 }
 
                 $listener = new LogoutListener(
-                    $app['security'],
+                    $app['security.token_storage'],
                     $app['security.http_utils'],
                     $app['security.authentication.logout_handler.'.$name],
                     $options,
@@ -476,7 +506,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
         $app['security.authentication_listener.switch_user._proto'] = $app->protect(function ($name, $options) use ($app, $that) {
             return $app->share(function () use ($app, $name, $options, $that) {
                 return new SwitchUserListener(
-                    $app['security'],
+                    $app['security.token_storage'],
                     $app['security.user_provider.'.$name],
                     $app['security.user_checker'],
                     $name,
@@ -524,7 +554,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
 
         if (isset($app['validator'])) {
             $app['security.validator.user_password_validator'] = $app->share(function ($app) {
-                return new UserPasswordValidator($app['security'], $app['security.encoder_factory']);
+                return new UserPasswordValidator($app['security.token_storage'], $app['security.encoder_factory']);
             });
 
             if (!isset($app['validator.validator_service_ids'])) {